// silap-colmirwcl/src/app/modules/silapapi/model/messaggio.ts
/**
 * 
 * 
 *
 * 
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { TipoMessaggio } from './tipoMessaggio';

export interface Messaggio { 
    codSilapDMessaggio?: string;
    dsSilapDMessaggio?: string;
    dsTitolo?: string;
    flgAttivo?: string;
    idSilapDMessaggio?: number;
    silapDTipoMessaggio?: TipoMessaggio;
}