// silap-colmirwcl/src/app/modules/silapapi/model/versamentoPvSospensione.ts
/**
 * 
 * 
 *
 * 
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { VersamentoMotivoSospensione } from './versamentoMotivoSospensione';

export interface VersamentoPvSospensione { 
    codUserAggiorn?: string;
    codUserInserim?: string;
    dAggiorn?: Date;
    dFineSospensione?: Date;
    dInizioSospensione?: Date;
    dInserim?: Date;
    esoDVersamentoMotivoSospensione?: VersamentoMotivoSospensione;
    idEsoTVersamentoPvSospensione?: number;
    nTimestamp?: number;
    numLavoratoriSospesi?: number;
    percSospensione?: number;
}